requirements.txt
Flask==3.0.3
mem0ai==0.0.18
ollama==0.3.1
openai==1.40.6